TrabajoFinal/reserva-citas/requirements.txt
fastapi
uvicorn
psycopg2-binary
sqlalchemy
requests
python-jose[cryptography]